requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
comtypes==1.1.7
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.2
django-crispy-forms==1.8.1
django-storages==1.9.1
djangorestframework==3.11.0
docx==0.2.4
docx2pdf==0.1.7
dropbox==9.4.0
gunicorn==20.0.4
idna==2.8
lxml==4.5.0
mysql==0.0.2
mysqlclient==1.4.6
oauthlib==3.1.0
Pillow==7.0.0
pipenv==2018.11.26
psycopg2==2.8.5
PyJWT==1.7.1
PyMySQL==0.9.3
python-docx==0.8.10
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
tqdm==4.46.0
urllib3==1.25.8
virtualenv==16.7.9
virtualenv-clone==0.5.3
whitenoise==5.1.0
